requirements.txt
django>=3.0,<3.1
django-extensions>=3.1,<4.0
django-filter>=2.2,<2.3
django-bootstrap4>=0.0,<1.2
django-cors-headers>=3.3,<3.4
djangorestframework>=3.12.2,<3.13
responses>=0.12,<1.0
requests>=2,<3
whitenoise>=5.2,<6.0
psycopg2-binary>=2.8,<2.9
gunicorn==20.0.4
